requirements.txt
fastapi
geopandas
pyyaml